requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-on-heroku==1.1.2
gunicorn==20.1.0
image==1.5.33
Pillow==9.0.1
psycopg2-binary==2.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
cloudinary==1.29.0
